requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.1
openai>=1.30.1
langchain>=0.1.16
langchain-openai>=0.1.6
langchain-experimental>=0.0.52
tiktoken>=0.7.0

ipykernel>=6.29.0